Add [storage_clients] as supported inventory section.
Allows limiting use of storage server to machines that require it.
A machine must be specified in [storage_clients] to use the storage server.
If the storage server is not specified and/or the machine is not listed in [storage_clients] provisioning of the machine will not use the storage server.

Signed-off-by: Joe Shimkus <jshimkus@redhat.com>