Generalize mount specifications.
This applies to both nfs and sshfs mounts.
Makes a mount specification a dictionary with keys 'server', 'src' and 'dest'.
'server' is the name/ip of the server.
'src' is the server-side path of the exported source.
'dest' is the client-side path where to mount the export.
Operationally the specifications are organized in a list.

Includes the new task import_user_nfs.yml in the role nfs_client.  This task creates the client-side mount point and mounts the export from the server.  These mounts are specified via the optional ansible extra variable 'user_nfs_mounts' which is a list of mount specifications as described previously.

Signed-off-by: Joe Shimkus <jshimkus@redhat.com>