Make certain loop device-derived VGs and LVs are active on cold boot.
As a last resort bunsen provisioning will use a loop device to provide
storage for /u1 and the test scratch device.  In the case of a cold
boot the current implementation leaves the VG and LVs created from the
loop device inactive.  This change adds a 'udevadm trigger' as part
of the systemd service starting the loop device which causes the VG and
LVs to be recognized and activated.