Fix bunsen identification of machine as a kvm guest.
The check that the machine's virtualization type is kvm is not sufficient.  A machine can have a virtualization type of kvm but not be a guest, rather a host providing virtualization.  In the case of nested virtualization the machine can be both.

This change uses relatively new ansible variables, if available, that disambiguate a machine's role as guest and/or host.  If these new variables are not available the change makes the determination the machine is a guest by its role not being host.

Signed-off-by: Joe Shimkus <jshimkus@redhat.com>